Refactor use `:fallback` option instead of `:fallback_to_devise`
The change is internal and does not affect how the gem behaves.

Strictly speaking, the `:fallback` option can now be used from
`handle_token_authentication_for` in place of `:fallback_to_devise`;
however, since it is still undocumented, there is no reason to
notice the extension and the gem behaviour can be considered unchanged.