Disable outperformed Cucumber scenario...
... related to setting configuration options from an initializer

Variables which behave as configuration  options can _sensu stricto_
always be modified from an initializer file. However, since most
files are required *before* the initializer files are read,
that ability is only effective when the variables are used at runtime.

The specs go a bit further than the scenarios, since not only
the `sign_in_token` option usage is tested, but also the usage of
the `header_names` option.

Both spec examples have in common a is very verbose setup; yet most
of what makes them interesting is that they cover a large part of the
authentication process, reducing the probablity of missing a call to
obsolete configuratio settings. There may be a better way to test that,
pointers are welcome!