Update dependencies
Just in case someone is wondering, the red Gemnasium and Hakiri badges
indicate that some Simple Token authentication dependencies are
out-of-date in the testing environment (i.e. the Travis CI build).
And that those out-fo-date dependencies eventually present security issues.

Running the test suite with slightly out-of-date dependencies does
not affect the security of the gem since there is no direct relation
between the versions of its dependencies in its own testing environment
and those which are in use in our production enviroments.

The red badges are, however, an alert for us, clients, to update quickly
all vulnerable dependencies in our own production environments!

    bundle update devise rack # for example